+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added support for nested subtasks
 - improved worktree pruning
+- workspace sidebar size is now stored on project level instead of task level
 
 ## [0.80.0]